Rename publication.php -> Publication.php, for consistency with class name.'
